requirements.txt
face_recognition==1.3.0
Flask==3.0.3
mysql_connector_repackaged==0.3.1
numpy==2.1.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
PyQt5==5.15.11
PyQt5_sip==12.15.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
